Extract fields helper into separate file
